+++ python_modules/dagster/dagster/core/definitions/repository.py
@@ -6,7 +6,7 @@ from .executable import ExecutableDefinition
 from .partition import PartitionScheduleDefinition, PartitionSetDefinition
 from .pipeline import PipelineDefinition
 from .schedule import ScheduleDefinition
-from .utils import check_for_invalid_name_and_warn
+from .utils import check_valid_name
 
 VALID_REPOSITORY_DATA_DICT_KEYS = {
     "pipelines",
@@ -511,7 +511,7 @@ class RepositoryDefinition(object):
     def __init__(
         self, name, repository_data, description=None,
     ):
-        self._name = check_for_invalid_name_and_warn(name)
+        self._name = check_valid_name(name)
         self._description = check.opt_str_param(description, "description")
         self._repository_data = check.inst_param(repository_data, "repository_data", RepositoryData)