+++ python_modules/dagster/dagster/core/definitions/schedule.py
@@ -11,7 +11,7 @@ from dagster.core.storage.tags import check_tags
 from dagster.utils import merge_dicts
 
 from .mode import DEFAULT_MODE_NAME
-from .utils import check_for_invalid_name_and_warn
+from .utils import check_valid_name
 
 
 class ScheduleExecutionContext(
@@ -106,7 +106,7 @@ class ScheduleDefinition(object):
         execution_timezone=None,
     ):
 
-        self._name = check_for_invalid_name_and_warn(name)
+        self._name = check_valid_name(name)
         self._cron_schedule = check.str_param(cron_schedule, "cron_schedule")
         self._pipeline_name = check.str_param(pipeline_name, "pipeline_name")
         self._run_config = check.opt_dict_param(run_config, "run_config")